requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2

langchain # core framework for building LLM applications
langchain-community # extra integrations and utilities for LangChain
faiss-cpu # fast vector database to store the embeddings
langchain-huggingface # connect huggingface models to perform embeddings 
langchain-text-splitters # to split the PDF text into smaller chunks for better processing
sentence-transformers # pre-trained models to convert chunks into vector embeddings
langchain-core # to handle documents, chains, of data etc ...